requirements.txt
transformers
diffusers
ctransformers
streamlit
accelerate
pypdf
peft
unstructured
tiktoken
faiss-cpu
python-magic
python-dotenv
langchain
langchain-community
langchain-openai
openai
requests
youtube_transcript_api
newspaper3k
lxml_html_clean
torch
pytesseract
pillow
python-docx
PyMuPDF
